ネクストの色を決める関数の作成。nextControllerができているが、次回nextManagerとnextBallManagerでファイル分けする。

+++ PuzzleBobble/program/BallController.h
@@ -9,7 +9,7 @@ private:
 	float speed = 0;
 	float rot = 0;
 	const float radius = 25.0f;
-	short state = 1;
+	int state = 1;
 
 private:
 	void BallBounce(float leftX,float rightX);
